Move whoami and introduce fetchHost in Conf
