+++ node/src/main/scala/coop/rchain/node/node.scala
@@ -3,57 +3,21 @@ package coop.rchain.node
 import coop.rchain.p2p, p2p.NetworkAddress, p2p.Network.KeysStore
 import coop.rchain.comm._, CommError._
 import com.typesafe.scalalogging.Logger
-import java.net.{InetAddress, NetworkInterface}
+import coop.rchain.node.repl._
 import java.util.UUID
 import java.io.File
-
-import scala.collection.JavaConverters._
 
 import cats._, cats.data._, cats.implicits._
 import coop.rchain.catscontrib._, Catscontrib._, ski._, TaskContrib._
 import monix.eval.Task
 
+import scala.concurrent.{ExecutionContext, Future}
+
 class NodeRuntime(conf: Conf) {
-
-  private val logger = Logger("main")
-  private def whoami(port: Int): Option[InetAddress] = {
-
-    val upnp = new UPnP(port)
-
-    logger.info(s"uPnP: ${upnp.localAddress} -> ${upnp.externalAddress}")
-
-    upnp.localAddress match {
-      case Some(addy) => Some(addy)
-      case None => {
-        val ifaces = NetworkInterface.getNetworkInterfaces.asScala.map(_.getInterfaceAddresses)
-        val addresses = ifaces
-          .flatMap(_.asScala)
-          .map(_.getAddress)
-          .toList
-          .groupBy(x => x.isLoopbackAddress || x.isLinkLocalAddress || x.isSiteLocalAddress)
-        if (addresses.contains(false)) {
-          Some(addresses(false).head)
-        } else {
-          val locals = addresses(true).groupBy(x => x.isLoopbackAddress || x.isLinkLocalAddress)
-          if (locals.contains(false)) {
-            Some(locals(false).head)
-          } else if (locals.contains(true)) {
-            Some(locals(true).head)
-          } else {
-            None
-          }
-        }
-      }
-    }
-  }
 
   import ApplicativeError_._
 
-  private val host =
-    conf.host.toOption match {
-      case Some(host) => host
-      case None       => whoami(conf.port()).fold("localhost")(_.getHostAddress)
-    }
+  private val host           = conf.fetchHost()
   private val name           = conf.name.toOption.fold(UUID.randomUUID.toString.replaceAll("-", ""))(id)
   private val address        = s"rnode://$name@$host:${conf.port()}"
   private val src            = p2p.NetworkAddress.parse(address).right.get
@@ -101,7 +65,7 @@ class NodeRuntime(conf: Conf) {
       http.stop
       net.broadcast(
         DisconnectMessage(ProtocolMessage.disconnect(net.local), System.currentTimeMillis))
-      logger.info("Goodbye.")
+      println("Goodbye.")
     }
   }
